vte: add api to set a manual defined palette
This patch allow the libtsm user to set a manually defined palette
on the tsm_vte object. The new API takes the palette as uint8_t *
data, and an int passing the number of colors defined. If that number
is different from the current number of colors used in libts, setting
the palette will fail. This means that adding new colors in libtsm is
not backward compatible with manually defined palettes.